+++ packages/context/src/lib/share-map/shared/share-map.service.ts
@@ -93,6 +93,9 @@ export class ShareMapService {
 
     const addedLayersQueryParamsWms = this.makeLayersByService(layers, contextLayersID, 'wms');
     const addedLayersQueryParamsWmts = this.makeLayersByService(layers, contextLayersID, 'wmts');
+    const addedLayersQueryParamsArcgisRest = this.makeLayersByService(layers, contextLayersID, 'arcgisrest');
+    const addedLayersQueryParamsImageArcgisRest = this.makeLayersByService(layers, contextLayersID, 'imagearcgisrest');
+    const addedLayersQueryParamsTileArcgisRest = this.makeLayersByService(layers, contextLayersID, 'tilearcgisrest');
 
     layersUrl = layersUrl.substr(0, layersUrl.length - 1);
 
@@ -110,7 +113,7 @@ export class ShareMapService {
       context = `${contextKey}=${this.contextService.context$.value.uri}`;
     }
 
-    let url = `${location.origin}${location.pathname}?${context}&${zoom}&${center}&${layersUrl}&${llc}&${addedLayersQueryParamsWms}&${llc}&${addedLayersQueryParamsWmts}`;
+    let url = `${location.origin}${location.pathname}?${context}&${zoom}&${center}&${layersUrl}&${llc}&${addedLayersQueryParamsWms}&${llc}&${addedLayersQueryParamsWmts}&${addedLayersQueryParamsArcgisRest}&${addedLayersQueryParamsImageArcgisRest}&${addedLayersQueryParamsTileArcgisRest}`;
 
     for (let i = 0; i < 5; i++) {
       url = url.replace(/&&/g, '&');
@@ -126,16 +129,20 @@ export class ShareMapService {
   private makeLayersByService(layers: Layer[], contextLayersID: any[], typeService: string): string {
 
     const addedLayersByService = [];
-    for (const layer of layers.filter(
-      l => l.dataSource.options && (l.dataSource.options.type === typeService)
-    )) {
+    for (const layer of layers.filter(l => l.dataSource.options?.type === typeService)) {
       if (contextLayersID.indexOf(layer.id) === -1) {
         const linkUrl = encodeURIComponent((layer.dataSource.options as any).url);
         let addedLayer = '';
-        if (layer.dataSource.options.type === 'wms') {
-          addedLayer = encodeURIComponent((layer.dataSource.options as any).params.LAYERS);
-        } else if (layer.dataSource.options.type === 'wmts') {
-          addedLayer = encodeURIComponent((layer.dataSource.options as any).layer);
+        switch (layer.dataSource.options.type.toLowerCase()) {
+          case 'wms':
+            addedLayer = encodeURIComponent((layer.dataSource.options as any).params.LAYERS);
+            break;
+          case 'wmts':
+          case 'arcgisrest':
+          case 'imagearcgisrest':
+          case 'tilearcgisrest':
+            addedLayer = encodeURIComponent((layer.dataSource.options as any).layer);
+            break;
         }
         const addedLayerPosition = `${addedLayer}:igoz${layer.zIndex}`;
 
@@ -158,10 +165,39 @@ export class ShareMapService {
 
     let addedLayersQueryParams = '';
     if (addedLayersByService.length >= 1) {
+      let linkUrlKey;
+      let layersKey;
+      /*
       const linkUrlKey = (typeService === 'wms') ? this.route.options.wmsUrlKey :
         (typeService === 'wmts') ? this.route.options.wmtsUrlKey : '' ;
       const layersKey = (typeService === 'wms') ? this.route.options.wmsLayersKey :
         (typeService === 'wmts') ? this.route.options.wmtsLayersKey : '' ;
+*/
+      switch (typeService.toLowerCase()) {
+        case 'wms':
+          linkUrlKey = this.route.options.wmsUrlKey;
+          layersKey = this.route.options.wmsLayersKey;
+          break;
+        case 'wmts':
+          linkUrlKey = this.route.options.wmtsUrlKey;
+          layersKey = this.route.options.wmtsLayersKey;
+          break;
+        case 'arcgisrest':
+          linkUrlKey = this.route.options.arcgisUrlKey;
+          layersKey = this.route.options.arcgisLayersKey;
+          break;
+        case 'imagearcgisrest':
+          linkUrlKey = this.route.options.iarcgisUrlKey;
+          layersKey = this.route.options.iarcgisLayersKey;
+          break;
+        case 'tilearcgisrest':
+          linkUrlKey = this.route.options.tarcgisUrlKey;
+          layersKey = this.route.options.tarcgisLayersKey;
+          break;
+        default:
+          linkUrlKey = '';
+          layersKey = '';
+      }
 
       let linkUrlQueryParams = '';
       let layersQueryParams = '';